# Better performance with sparse columns in aggregate functions