@CurtizJ 2022-05-28
None of the related tests was not run, because of randomization, I think.

![Screenshot 2022-05-28 at 20 43 30](https://user-images.githubusercontent.com/20361854/170838825-2b135111-f4a0-409b-9446-c60f38792d75.png)


Here is the result of manual testing on table with columns with 99% ratio of defaults:

```
clickhouse benchmark <<< "SELECT sum(u64) FROM test_sparse FORMAT Null" --cumulative

Old:
localhost:9000, queries 237, QPS: 50.411, RPS: 5041054589.999, MiB/s: 770.766, result RPS: 0.000, result MiB/s: 0.000.

0.000%          0.018 sec.
10.000%         0.019 sec.
20.000%         0.019 sec.
30.000%         0.019 sec.
40.000%         0.019 sec.
50.000%         0.019 sec.
60.000%         0.019 sec.
70.000%         0.020 sec.
80.000%         0.020 sec.
90.000%         0.020 sec.
95.000%         0.021 sec.
99.000%         0.022 sec.
99.900%         0.043 sec.
99.990%         0.043 sec.

New:
localhost:9000, queries 654, QPS: 146.873, RPS: 14687315920.448, MiB/s: 2245.658, result RPS: 0.000, result MiB/s: 0.000.

0.000%          0.006 sec.
10.000%         0.007 sec.
20.000%         0.007 sec.
30.000%         0.007 sec.
40.000%         0.007 sec.
50.000%         0.007 sec.
60.000%         0.007 sec.
70.000%         0.007 sec.
80.000%         0.007 sec.
90.000%         0.007 sec.
95.000%         0.007 sec.
99.000%         0.008 sec.
99.900%         0.025 sec.
99.990%         0.028 sec.
```

```
clickhouse benchmark <<< "SELECT sum(u64) FROM test_sparse GROUP BY id % 11 FORMAT Null" --cumulative

Old:
localhost:9000, queries 94, QPS: 10.548, RPS: 1054816226.316, MiB/s: 8208.969, result RPS: 0.000, result MiB/s: 0.000.

0.000%          0.089 sec.
10.000%         0.093 sec.
20.000%         0.093 sec.
30.000%         0.093 sec.
40.000%         0.093 sec.
50.000%         0.094 sec.
60.000%         0.094 sec.
70.000%         0.094 sec.
80.000%         0.096 sec.
90.000%         0.097 sec.
95.000%         0.097 sec.
99.000%         0.115 sec.
99.900%         0.131 sec.
99.990%         0.131 sec.

New:
localhost:9000, queries 130, QPS: 13.117, RPS: 1311701499.934, MiB/s: 10208.145, result RPS: 0.000, result MiB/s: 0.000.

0.000%          0.072 sec.
10.000%         0.075 sec.
20.000%         0.075 sec.
30.000%         0.075 sec.
40.000%         0.075 sec.
50.000%         0.075 sec.
60.000%         0.075 sec.
70.000%         0.076 sec.
80.000%         0.078 sec.
90.000%         0.078 sec.
95.000%         0.080 sec.
99.000%         0.095 sec.
99.900%         0.100 sec.
99.990%         0.100 sec.
```
```
clickhouse benchmark <<< "SELECT uniq(str) FROM test_sparse FORMAT Null" --cumulative

Old:
localhost:9000, queries 119, QPS: 11.026, RPS: 1102640824.032, MiB/s: 852.124, result RPS: 0.000, result MiB/s: 0.000.

0.000%          0.085 sec.
10.000%         0.087 sec.
20.000%         0.088 sec.
30.000%         0.089 sec.
40.000%         0.089 sec.
50.000%         0.089 sec.
60.000%         0.090 sec.
70.000%         0.091 sec.
80.000%         0.093 sec.
90.000%         0.096 sec.
95.000%         0.098 sec.
99.000%         0.102 sec.
99.900%         0.114 sec.
99.990%         0.114 sec.

New:
localhost:9000, queries 353, QPS: 35.685, RPS: 3568540593.840, MiB/s: 2757.778, result RPS: 0.000, result MiB/s: 0.000.

0.000%          0.026 sec.
10.000%         0.027 sec.
20.000%         0.027 sec.
30.000%         0.027 sec.
40.000%         0.028 sec.
50.000%         0.028 sec.
60.000%         0.028 sec.
70.000%         0.028 sec.
80.000%         0.028 sec.
90.000%         0.028 sec.
95.000%         0.029 sec.
99.000%         0.036 sec.
99.900%         0.073 sec.
99.990%         0.073 sec.
```